+++ prj/NAppCompilers.cmake
@@ -144,6 +144,7 @@ elseif (${CMAKE_SYSTEM_NAME} STREQUAL "Darwin")
     # In CMake 4, CMAKE_OSX_SYSROOT is empty by default
     if (NOT OSX_SYSROOT)
         execute_process(COMMAND xcrun --sdk macosx --show-sdk-path OUTPUT_VARIABLE OSX_SYSROOT)
+        string(REPLACE "\n" "" OSX_SYSROOT ${OSX_SYSROOT})
     endif()
 
     if (NOT OSX_SYSROOT)
@@ -153,7 +154,6 @@ elseif (${CMAKE_SYSTEM_NAME} STREQUAL "Darwin")
     get_filename_component(CMAKE_BASE_OSX_SDK ${OSX_SYSROOT} NAME)
     string(REPLACE "MacOSX" "" CMAKE_BASE_OSX_SDK ${CMAKE_BASE_OSX_SDK})
     string(REPLACE ".sdk" "" CMAKE_BASE_OSX_SDK ${CMAKE_BASE_OSX_SDK})
-    string(REPLACE "\n" "" CMAKE_BASE_OSX_SDK ${CMAKE_BASE_OSX_SDK})
 
     # Get the Deployment target
     set(CMAKE_OSX_DEPLOYMENT_TARGET ${CMAKE_BASE_OSX_SDK} CACHE STRING "Minimun macOS SDK to support")